Fixes here and there
[compo] Give new namespace to ECS code
[comp_app] Make sure loading happens in prioritized order
[gleam] Fix some OpenGL 4.1 compatibility issues

+++ src/peripheral/include/peripherals/libc/memory_ops.h
@@ -107,7 +107,7 @@ using namespace ::libc_types;
 FORCEDINLINE u32 page_size()
 {
 #if defined(COFFEE_UNIXPLAT)
-    int pagesize = ::sysconf(_SC_PAGESIZE);
+    i64 pagesize = ::sysconf(_SC_PAGESIZE);
 
     return C_FCAST<u32>(pagesize > 0 ? pagesize : 4_kB);
 #else
